refactor(tests): move PT wrapper tests into the squad mirror, split by concern
tests/test_squad_penetration_tester.py (488) was the last top-level
test_squad_<agent>.py for an agent that already has a tests/squad/<agent>/
mirror (disclosure_coordinator and programme_manager already migrated
their wrapper tests to tests/squad/<agent>/test_tools.py). Move it into
the mirror and split by concern, no coverage change (60 tests, none
dropped):

  test_tools.py          <- TestPenetrationTesterTools: the bespoke probe
                            / recon / findings wrapper tests (each mocks
                            the specific helper it forwards to)
  test_cloud_wrappers.py <- the FQDN / endpoint cloud-wrapper parametrize
                            tables + their closed-world drift checks

Largest resulting file is test_cloud_wrappers.py at 355 lines, under the
450 bar.
